+++ __tests__/utils/AMSClientUtils.web.spec.ts
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 
-import { whitelistedUrls, isUrlWhitelisted, shouldUseFramedMode, retrieveRegionBasedUrl } from "../../src/utils/AMSClientUtils";
+import { whitelistedUrls, isUrlWhitelisted, shouldUseFramedMode, regionBasedSupportedUrls, regionBasedUrlSupportedVersions, isRegionBasedUrlSupported, retrieveRegionBasedUrl } from "../../src/utils/AMSClientUtils";
 
 describe("AMSClientUtils", () => {
     const { location } = window;
@@ -60,6 +60,27 @@ describe("AMSClientUtils", () => {
         expect(shouldUseFramedMode(true)).toBe(true);
     });
 
+    it("isRegionBasedUrlSupported() should return 'true' if URL is supported", () => {
+        for (const url of regionBasedSupportedUrls) {
+            const isSupported = isRegionBasedUrlSupported(url);
+            expect(isSupported).toBe(true);
+        }
+    });
+
+    it("isRegionBasedUrlSupported() should return 'false' if URL is not supported", () => {
+        const unsupportedUrl = [
+            "https://ocprodocprodnamgs.blob.core.usgovcloudapi.net",
+            "https://microsoft.com",
+            "https://bing.com",
+            "https://contoso.com"
+        ];
+
+        for (const url of unsupportedUrl) {
+            const isSupported = isRegionBasedUrlSupported(url);
+            expect(isSupported).toBe(false);
+        }
+    });
+
     it("retrieveRegionBasedUrl() with '' as widget snippet base url should return ''", () => {
         const widgetSnippetBaseUrl = "";
         const regionBasedUrl = retrieveRegionBasedUrl(widgetSnippetBaseUrl);
@@ -67,10 +88,48 @@ describe("AMSClientUtils", () => {
     });
 
     it("retrieveRegionBasedUrl() should return region based url accordingly", () => {
-        for (const url of whitelistedUrls) {
+        for (const url of regionBasedSupportedUrls) {
             const widgetSnippetBaseUrl = url;
             const regionBasedUrl = retrieveRegionBasedUrl(widgetSnippetBaseUrl);
             expect(regionBasedUrl).toBe(`${widgetSnippetBaseUrl}/livechatwidget/v2scripts/ams`);
+        }
+    });
+
+    it("retrieveRegionBasedUrl() should return '' if widget snippet base url is not supported", () => {
+        const unsupportedUrl = [
+            "https://ocprodocprodnamgs.blob.core.usgovcloudapi.net",
+            "https://microsoft.com",
+            "https://bing.com",
+            "https://contoso.com"
+        ];
+
+        for (const url of unsupportedUrl) {
+            const regionBasedUrl = retrieveRegionBasedUrl(url);
+            expect(regionBasedUrl).toBe("");
+        }
+    });
+
+    it("retrieveRegionBasedUrl() should NOT return '' if version is supported", () => {
+        const url = regionBasedSupportedUrls[0];
+        for (const version of regionBasedUrlSupportedVersions) {
+            jest.replaceProperty(require('@microsoft/omnichannel-amsclient/package.json'), 'version', version);
+            const regionBasedUrl = retrieveRegionBasedUrl(url);
+            expect(regionBasedUrl).not.toBe("");
+        }
+    });
+
+    it("retrieveRegionBasedUrl() should return '' if version is not supported", () => {
+        const url = regionBasedSupportedUrls[0];
+        const unsupportedVersions = [
+            "0.1.5",
+            "0.1.6",
+            "0.1.7",
+        ];
+
+        for (const version of unsupportedVersions) {
+            jest.replaceProperty(require('@microsoft/omnichannel-amsclient/package.json'), 'version', version);
+            const regionBasedUrl = retrieveRegionBasedUrl(url);
+            expect(regionBasedUrl).toBe("");
         }
     });
 });